+++ docs/CHANGELOG.md
@@ -76,6 +76,6 @@ v1.5.1
 - Add scatter unit tests and fix scatter imports
 - Fix black renders due to non-hidden particle emitter
 
-
 v1.6.0
 - Add geometric tile pattern materials
+- Tune window parameters and materials